requirements.txt
Flask==2.3.3
pymongo==4.6.1
python-dotenv==1.0.0
WTForms==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
bcrypt==4.1.2
Flask-Mail==0.9.1
python-docx==1.1.0
PyPDF2==3.0.1
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
python-magic==0.4.27
Pillow==10.1.0
dnspython==2.7.0 